Clean up of package `stream`, pt1
