+++ core/src/main/java/org/frankframework/stream/document/JsonEventHandler.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package org.frankframework.stream;
+package org.frankframework.stream.document;
 
 import org.xml.sax.SAXException;
