+++ core/src/main/java/org/frankframework/stream/document/xml/JsonXslt3XmlHandler.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package org.frankframework.stream.xml;
+package org.frankframework.stream.document.xml;
 
 import java.io.IOException;
 
@@ -21,7 +21,7 @@ import jakarta.json.stream.JsonParser;
 import jakarta.json.stream.JsonParser.Event;
 import lombok.Getter;
 import lombok.Setter;
-import org.frankframework.stream.JsonEventHandler;
+import org.frankframework.stream.document.JsonEventHandler;
 import org.xml.sax.ContentHandler;
 import org.xml.sax.SAXException;
 import org.xml.sax.helpers.AttributesImpl;
