+++ core/src/main/java/org/frankframework/threading/ThreadLifeCycleEventListener.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package org.frankframework.stream;
+package org.frankframework.threading;
 
 public interface ThreadLifeCycleEventListener<T> {
